requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
notebook>=6.0.0
streamlit>=1.0.0
pillow>=7.1.0
